move generateUUID into utility function.
