+++ src/editors/uuid.js
@@ -1,3 +1,4 @@
+import { generateUUID } from '../utilities.js'
 import { StringEditor } from './string.js'
 
 export class UuidEditor extends StringEditor {
@@ -36,18 +37,7 @@ export class UuidEditor extends StringEditor {
   }
 
   getUuid () {
-    /* https://stackoverflow.com/questions/105034/create-guid-uuid-in-javascript */
-    let d = new Date().getTime()
-
-    if (typeof performance !== 'undefined' && typeof performance.now === 'function') {
-      d += performance.now() /* use high-precision timer if available */
-    }
-
-    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-      const r = (d + Math.random() * 16) % 16 | 0
-      d = Math.floor(d / 16)
-      return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16)
-    })
+    return generateUUID()
   }
 
   testUuid (value) {
