+++ src/utilities.js
@@ -86,3 +86,21 @@ export function isInteger (value) {
   const v = parseInt(value)
   return match !== null && !isNaN(v) && isFinite(v)
 }
+
+/* This function generates a uuid.
+https://stackoverflow.com/questions/105034/create-guid-uuid-in-javascript
+TODO: It will be probably better to move to: https://www.npmjs.com/package/uuid
+*/
+export function generateUUID () {
+  let d = new Date().getTime()
+
+  if (typeof performance !== 'undefined' && typeof performance.now === 'function') {
+    d += performance.now() /* use high-precision timer if available */
+  }
+
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+    const r = (d + Math.random() * 16) % 16 | 0
+    d = Math.floor(d / 16)
+    return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16)
+  })
+}
